requirements.txt
openai
python-dotenv
groq
phidata
packaging
duckduckgo-search
uvicorn
fastapi
streamlit
google-generativeai

